Clean Up (Eclipse built in)
This will add a lot of @Override, fix imports, etc

+++ CommitSummarizer/CommitSummarizer.core/src/main/java/co/edu/unal/colswe/changescribe/core/stereotype/taxonomy/TypeStereotypeRules.java
@@ -73,7 +73,7 @@ public class TypeStereotypeRules
     
     protected boolean checkForEntity() {
         final boolean rule1 = !this.accessorMethods.isEmpty() && !this.mutatorMethods.isEmpty();
-        final boolean rule2 = 0.4 * this.totalMethods.size() <= (double)this.collaborationalMethods.size() && (double)this.collaborationalMethods.size() <= 0.6 * this.totalMethods.size();
+        final boolean rule2 = 0.4 * this.totalMethods.size() <= this.collaborationalMethods.size() && this.collaborationalMethods.size() <= 0.6 * this.totalMethods.size();
         final boolean rule3 = this.controllerMethods.isEmpty();
         return rule1 && rule2 && rule3;
     }
@@ -90,37 +90,37 @@ public class TypeStereotypeRules
     }
     
     protected boolean checkForDataProvider() {
-        final boolean rule1 = (double)this.accessorMethods.size() > 2.0 * this.mutatorMethods.size();
+        final boolean rule1 = this.accessorMethods.size() > 2.0 * this.mutatorMethods.size();
         final Set<StereotypedMethod> temp = this.controllerMethods;
         temp.addAll(this.factoryMethods);
-        final boolean rule2 = (double)this.accessorMethods.size() > 2.0 * temp.size();
+        final boolean rule2 = this.accessorMethods.size() > 2.0 * temp.size();
         return rule1 && rule2;
     }
     
     protected boolean checkForCommander() {
-        final boolean rule1 = (double)this.mutatorMethods.size() > 2.0 * this.accessorMethods.size();
+        final boolean rule1 = this.mutatorMethods.size() > 2.0 * this.accessorMethods.size();
         final Set<StereotypedMethod> temp = this.controllerMethods;
         temp.addAll(this.factoryMethods);
-        final boolean rule2 = (double)this.mutatorMethods.size() > 2.0 * temp.size();
+        final boolean rule2 = this.mutatorMethods.size() > 2.0 * temp.size();
         return rule1 && rule2;
     }
     
     protected boolean checkForBoundary() {
         final boolean rule1 = this.collaboratorMethods.size() >= this.totalMethods.size() - this.collaboratorMethods.size();
-        final boolean rule2 = (double)this.factoryMethods.size() <= 0.5 * this.totalMethods.size();
-        final boolean rule3 = (double)this.controllerMethods.size() <= 0.3333333333333333 * this.totalMethods.size();
+        final boolean rule2 = this.factoryMethods.size() <= 0.5 * this.totalMethods.size();
+        final boolean rule3 = this.controllerMethods.size() <= 0.3333333333333333 * this.totalMethods.size();
         return rule1 && rule2 && rule3;
     }
     
     protected boolean checkForFactory() {
-        final boolean rule1 = (double)this.factoryMethods.size() >= 0.6666666666666666 * this.totalMethods.size();
+        final boolean rule1 = this.factoryMethods.size() >= 0.6666666666666666 * this.totalMethods.size();
         return rule1;
     }
     
     protected boolean checkForController() {
         final HashSet<StereotypedMethod> temp = new HashSet<StereotypedMethod>(this.collaborationalMethods);
         temp.addAll(this.factoryMethods);
-        final boolean rule1 = (double)temp.size() >= 0.6666666666666666 * this.totalMethods.size();
+        final boolean rule1 = temp.size() >= 0.6666666666666666 * this.totalMethods.size();
         final boolean rule2 = !this.controllerMethods.isEmpty() || !this.factoryMethods.isEmpty();
         final boolean rule3 = !this.mutatorMethods.isEmpty() || !this.accessorMethods.isEmpty();
         return rule1 && rule2 && rule3;
@@ -138,32 +138,32 @@ public class TypeStereotypeRules
     }
     
     protected boolean checkForLargeClass() {
-        final boolean rule1 = 0.2 * this.totalMethods.size() <= (double)(this.accessorMethods.size() + this.mutatorMethods.size()) && (double)(this.accessorMethods.size() + this.mutatorMethods.size()) <= 0.6666666666666666 * this.totalMethods.size();
+        final boolean rule1 = 0.2 * this.totalMethods.size() <= this.accessorMethods.size() + this.mutatorMethods.size() && this.accessorMethods.size() + this.mutatorMethods.size() <= 0.6666666666666666 * this.totalMethods.size();
         final HashSet<StereotypedMethod> temp = new HashSet<StereotypedMethod>(this.controllerMethods);
         temp.addAll(this.factoryMethods);
-        final boolean rule2 = 0.2 * this.totalMethods.size() <= (double)temp.size() && (double)temp.size() <= 0.6666666666666666 * this.totalMethods.size();
+        final boolean rule2 = 0.2 * this.totalMethods.size() <= temp.size() && temp.size() <= 0.6666666666666666 * this.totalMethods.size();
         final boolean rule3 = !this.controllerMethods.isEmpty();
         final boolean rule4 = !this.factoryMethods.isEmpty();
         final boolean rule5 = !this.accessorMethods.isEmpty();
         final boolean rule6 = !this.mutatorMethods.isEmpty();
-        final boolean rule7 = (double)this.totalMethods.size() > this.methodsMean + this.methodsStdDev;
+        final boolean rule7 = this.totalMethods.size() > this.methodsMean + this.methodsStdDev;
         return rule1 && rule2 && rule3 && rule4 && rule5 && rule6 && rule7;
     }
     
     protected boolean checkForLazyClass() {
-        final boolean rule1 = (double)this.incidentalMethods.size() >= 0.3333333333333333 * this.totalMethods.size();
-        final boolean rule2 = (double)(this.getMethods.size() + this.setMethods.size() + this.incidentalMethods.size()) >= 0.8 * this.totalMethods.size();
+        final boolean rule1 = this.incidentalMethods.size() >= 0.3333333333333333 * this.totalMethods.size();
+        final boolean rule2 = this.getMethods.size() + this.setMethods.size() + this.incidentalMethods.size() >= 0.8 * this.totalMethods.size();
         return rule1 && rule2;
     }
     
     protected boolean checkForDegenerate() {
-        final boolean rule1 = (double)this.emptyMethods.size() >= 0.3333333333333333 * this.totalMethods.size();
-        final boolean rule2 = (double)(this.getMethods.size() + this.setMethods.size() + this.emptyMethods.size()) >= 0.8 * this.totalMethods.size();
+        final boolean rule1 = this.emptyMethods.size() >= 0.3333333333333333 * this.totalMethods.size();
+        final boolean rule2 = this.getMethods.size() + this.setMethods.size() + this.emptyMethods.size() >= 0.8 * this.totalMethods.size();
         return rule1 && rule2;
     }
     
     protected boolean checkForDataClass() {
-        final boolean rule1 = (double)(this.getMethods.size() + this.setMethods.size()) > 0.0;
+        final boolean rule1 = this.getMethods.size() + this.setMethods.size() > 0.0;
         final boolean rule2 = this.totalMethods.size() == this.getMethods.size() + this.setMethods.size();
         final boolean rule3 = this.collaborationalMethods.isEmpty();
         return rule1 && rule2 && rule3;
